move predictor base class to module
